:bug: Fix import web handling
