+++ lib/domain/repository/game_state_repository_mobile.dart
@@ -1,0 +1,15 @@
+import 'package:wordlef/domain/model/game_saved_state.dart';
+
+import 'game_state_repository.dart';
+
+class GameStateRepository extends IGameStateRepository {
+  @override
+  void autoSave(OnInactiveGameCallback onInactiveGame) {
+    // Nothing
+  }
+
+  @override
+  GameSavedState? restore() {
+    return null;
+  }
+}
